Implement an interface for GLX server modules.
Added a new directory, glxserver, which implements a libglx.so module for the X
server. The module defines an interface where drivers can load GLX
implementations, and assign them to individual screens.

The GLVND module then takes each request and forwards it to whichever vendor
is assigned to a screen.