GLX: Always try to look up a screen number for a drawable.
In __glXVendorFromDrawable, send the GLXQueryDrawable request if GLX is
supported, instead of requiring that GLX_EXT_libglvnd be supported on every
screen.

If the server doesn't support GLX_EXT_libglvnd, then it won't send back a
screen number, and libGLX.so will default to screen zero like it does now.