glxserver: Simplify initialization and registration.
Use only a single callback for registering vendor libraries.

Drivers can now call __glXCreateVendor at any time. It takes a pointer to a
__GLXserverImports struct, so we don't have to worry about uninitialized vendor
handles.

The idea is that a driver can create a vendor handle any time in or before
InitExtensions, and then in the InitExtensions callback, the driver can assign
the vendor to any screens.

Calling __glXSetScreenVendor on a screen that already has a vendor is now an
error.

Drivers can (optionally) destroy the vendor handle from the
__GLXserverImports::extensionCloseDown callback.