glxserver: Rework the context tag private data functions.
Add separate functions to set and look up the private data pointer for a
context tag.

Vendors now use the new functions, getContextTagPrivate and
setContextTagPrivate to manage the private data pointer for a context tag.

__GLXserverImports::makeCurrent no longer takes the old and new private data
pointers as arguments. Instead, the vendor can use getContextTagPrivate to get
the old pointer, and setContextTagPrivate to set the new pointer.