GLX: Added a GLVNDappProfileVendor::onlyInServerList flag.
If the flag is set, then libglvnd will check the GLX_VENDOR_NAMES_EXT string
for the default screen. If the vendor's name isn't listed, then libglvnd won't
try to load the vendor for GPU offloading.

This is an optional optimization. If a vendor library relies on a corresponding
module in the server, then this provides a simple way to check for that module
without having to load the vendor first.

We'll probably want to define a new string for this rather than using
GLX_VENDOR_NAMES_EXT.