Load app profile data from a config file.
Libglvnd will now look for config files to load the app profile data. The
config files are JSON-based, and should allow for configuring applications
both in terms of specific vendors (e.g., "mesa") or general rules (e.g.,
"performance").

The config file parsing is largely isolated, so the rest of libglvnd still only
has to care about a list of (vendors name, vendor data) tuples. That also means
that it'll be easy to change how libglvnd loads profiles later on if needed.